Assert payment has source during processing.
This security fix prevents users from completing an order without processing
credit card data by gateway. It was possible by tinkering with the payment form
and removing the source_attributes fieldsets. An easy task on modern browsers
that provide DOM inspectors.

It's not critical. All the user is able to do without this patch is completing
an order similarly to if the store provided "Check" payment method. The order
will appear in completed orders list in admin panel with "Credit Card"
payment method and "checkout" payment state. An exception would be thrown
if store owner tried editing the payment.